Cambios en pantallas y correciones de componentes
