+++ src/components/SliderComponent/index.js
@@ -1,0 +1,3 @@
+import SliderComponent from './SliderComponent'
+
+export default SliderComponent
