+++ src/components/SliderComponent/sliderComponent.js
@@ -1,31 +1,28 @@
-/* eslint-disable react/destructuring-assignment */
-import React from 'react'
+import React, { useState } from 'react'
 import { Slider } from '@miblanchard/react-native-slider'
-import {
-  AppRegistry, View, Text,
-} from 'react-native'
+import { View, Text } from 'react-native'
 import styles from './styles'
 
-// eslint-disable-next-line import/prefer-default-export
-export class SliderComponent extends React.Component {
-  state = {
-    value: 20,
+const SliderComponent = () => {
+  const [value, setValue] = useState(20)
+
+  const handleValueChange = newValue => {
+    setValue(newValue)
   }
 
-  render() {
-    return (
-      <View style={styles.container}>
-        <Slider
-          value={this.state.value}
-          onValueChange={value => this.setState({ value })}
-        />
-        <Text>
-          Value:
-          {this.state.value}
-        </Text>
-      </View>
-    )
-  }
+  return (
+    <View style={styles.container}>
+      <Slider
+        value={value}
+        onValueChange={handleValueChange}
+      />
+      <Text>
+        Value:
+        {' '}
+        {value}
+      </Text>
+    </View>
+  )
 }
 
-AppRegistry.registerComponent('SliderComponent', () => SliderComponent)
+export default SliderComponent
