+++ src/components/TextCard/index.js
@@ -1,0 +1,3 @@
+import TextCard from './TextCard'
+
+export default TextCard
