+++ src/components/TextCard/styles.js
@@ -1,18 +1,3 @@
-import React from 'react'
-import { View, Text, StyleSheet } from 'react-native'
-import * as Font from 'expo-font'
-import SamsungOne from '../fonts/SamsungOne-400.ttf'
-
-const TextCard = ({ number }) => (
-  <View style={styles.container}>
-    <Text style={styles.text}>
-      {' '}
-      Total de aves:
-      {number}
-    </Text>
-  </View>
-)
-
 const styles = StyleSheet.create({
   container: {
     alignItems: 'center',
@@ -36,4 +21,4 @@ const styles = StyleSheet.create({
   },
 })
 
-export default TextCard
+export default styles
