+++ src/components/TrafficLight/index.js
@@ -1,0 +1,3 @@
+import TrafficLight from './TrafficLight'
+
+export default TrafficLight
