+++ src/components/TrafficLight/styles.js
@@ -1,24 +1,3 @@
-import React from 'react'
-import { View, Text, StyleSheet } from 'react-native'
-import * as Font from 'expo-font'
-import SamsungOne from '../fonts/SamsungOne-400.ttf'
-
-const TrafficLight = ({ topValue, middleValue, bottomValue }) => (
-  <View style={styles.container}>
-    <Text style={styles.boxHeaderText}>C.A.</Text>
-    <View style={[styles.box, styles.greenBox]}>
-      <Text style={styles.boxText}>{topValue}</Text>
-
-    </View>
-    <View style={[styles.box, styles.orangeBox]}>
-      <Text style={styles.boxText}>{middleValue}</Text>
-    </View>
-    <View style={[styles.box, styles.redBox]}>
-      <Text style={styles.boxText}>{bottomValue}</Text>
-    </View>
-  </View>
-)
-
 const styles = StyleSheet.create({
   box: {
     alignItems: 'center',
@@ -47,7 +26,6 @@ const styles = StyleSheet.create({
   container: {
     alignItems: 'center',
     backgroundColor: '#FFFFFF',
-    backgroundColor: '#FFFFFF',
     borderRadius: 5,
     elevation: 4,
     marginRight: 15,
@@ -69,5 +47,4 @@ const styles = StyleSheet.create({
     backgroundColor: 'red',
   },
 })
-
-export default TrafficLight
+export default styles
