+++ src/screens/HomeWorkerScreen/HomeWorkerScreen.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { View, ScrollView, StatusBar } from 'react-native'
 import PropTypes from 'prop-types'
 import { CardGalera } from '../../components'
-import TextCard from '../../components/TextCard'
+import TextCard from '../../components/TextCard/TextCard'
 import useApi from '../../hooks/useApi/useApi'
 
 const HomeWorkerScreen = ({ navigation }) => {
