+++ src/screens/index.js
@@ -2,10 +2,14 @@ import ChoiceScreen from './ChoiceScreen'
 import CreationScreen from './CreationScreen'
 import DetailsScreen from './DetailsScreen'
 import HomeWorkerScreen from './HomeWorkerScreen'
+import LoginAdministrator from './LoginAdministrator'
+import InitialScreen from './InitialScreen'
 
 export {
   ChoiceScreen,
   CreationScreen,
   DetailsScreen,
   HomeWorkerScreen,
+  LoginAdministrator,
+  InitialScreen
 }
